Prepare repository to be merged into core repos
