--- rapidcore.redis.sln
@@ -1,67 +1,0 @@
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.26430.13
-MinimumVisualStudioVersion = 15.0.26124.0
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "rapidcore.redis", "src\rapidcore.redis.csproj", "{BA9A80CA-8E71-4BB4-AC3D-4FE8542ED429}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{532C8B26-92F2-4EC6-A483-6346CC0692B8}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "unittests", "test\unit\unittests.csproj", "{6F41AD23-5478-4C7B-B710-4604C2447AE5}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "functionaltest", "test\functional\functionaltest.csproj", "{587743BB-C561-434C-8C40-94A2E8194C3F}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|Any CPU = Release|Any CPU
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{BA9A80CA-8E71-4BB4-AC3D-4FE8542ED429}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{BA9A80CA-8E71-4BB4-AC3D-4FE8542ED429}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{BA9A80CA-8E71-4BB4-AC3D-4FE8542ED429}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{BA9A80CA-8E71-4BB4-AC3D-4FE8542ED429}.Debug|x64.Build.0 = Debug|Any CPU
-		{BA9A80CA-8E71-4BB4-AC3D-4FE8542ED429}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{BA9A80CA-8E71-4BB4-AC3D-4FE8542ED429}.Debug|x86.Build.0 = Debug|Any CPU
-		{BA9A80CA-8E71-4BB4-AC3D-4FE8542ED429}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{BA9A80CA-8E71-4BB4-AC3D-4FE8542ED429}.Release|Any CPU.Build.0 = Release|Any CPU
-		{BA9A80CA-8E71-4BB4-AC3D-4FE8542ED429}.Release|x64.ActiveCfg = Release|Any CPU
-		{BA9A80CA-8E71-4BB4-AC3D-4FE8542ED429}.Release|x64.Build.0 = Release|Any CPU
-		{BA9A80CA-8E71-4BB4-AC3D-4FE8542ED429}.Release|x86.ActiveCfg = Release|Any CPU
-		{BA9A80CA-8E71-4BB4-AC3D-4FE8542ED429}.Release|x86.Build.0 = Release|Any CPU
-		{6F41AD23-5478-4C7B-B710-4604C2447AE5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{6F41AD23-5478-4C7B-B710-4604C2447AE5}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{6F41AD23-5478-4C7B-B710-4604C2447AE5}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{6F41AD23-5478-4C7B-B710-4604C2447AE5}.Debug|x64.Build.0 = Debug|Any CPU
-		{6F41AD23-5478-4C7B-B710-4604C2447AE5}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{6F41AD23-5478-4C7B-B710-4604C2447AE5}.Debug|x86.Build.0 = Debug|Any CPU
-		{6F41AD23-5478-4C7B-B710-4604C2447AE5}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{6F41AD23-5478-4C7B-B710-4604C2447AE5}.Release|Any CPU.Build.0 = Release|Any CPU
-		{6F41AD23-5478-4C7B-B710-4604C2447AE5}.Release|x64.ActiveCfg = Release|Any CPU
-		{6F41AD23-5478-4C7B-B710-4604C2447AE5}.Release|x64.Build.0 = Release|Any CPU
-		{6F41AD23-5478-4C7B-B710-4604C2447AE5}.Release|x86.ActiveCfg = Release|Any CPU
-		{6F41AD23-5478-4C7B-B710-4604C2447AE5}.Release|x86.Build.0 = Release|Any CPU
-		{587743BB-C561-434C-8C40-94A2E8194C3F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{587743BB-C561-434C-8C40-94A2E8194C3F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{587743BB-C561-434C-8C40-94A2E8194C3F}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{587743BB-C561-434C-8C40-94A2E8194C3F}.Debug|x64.Build.0 = Debug|Any CPU
-		{587743BB-C561-434C-8C40-94A2E8194C3F}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{587743BB-C561-434C-8C40-94A2E8194C3F}.Debug|x86.Build.0 = Debug|Any CPU
-		{587743BB-C561-434C-8C40-94A2E8194C3F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{587743BB-C561-434C-8C40-94A2E8194C3F}.Release|Any CPU.Build.0 = Release|Any CPU
-		{587743BB-C561-434C-8C40-94A2E8194C3F}.Release|x64.ActiveCfg = Release|Any CPU
-		{587743BB-C561-434C-8C40-94A2E8194C3F}.Release|x64.Build.0 = Release|Any CPU
-		{587743BB-C561-434C-8C40-94A2E8194C3F}.Release|x86.ActiveCfg = Release|Any CPU
-		{587743BB-C561-434C-8C40-94A2E8194C3F}.Release|x86.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(NestedProjects) = preSolution
-		{6F41AD23-5478-4C7B-B710-4604C2447AE5} = {532C8B26-92F2-4EC6-A483-6346CC0692B8}
-		{587743BB-C561-434C-8C40-94A2E8194C3F} = {532C8B26-92F2-4EC6-A483-6346CC0692B8}
-	EndGlobalSection
-EndGlobal
