+++ src/redis/.travis.yml
