+++ src/redis/test-functional/functionaltest.csproj
@@ -1,16 +1,17 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <TargetFramework>netcoreapp1.1</TargetFramework>
-    <RootNamespace>RapidCore.Redis.UnitTests</RootNamespace>
-    <AssemblyName>RapidCore.Redis.UnitTests</AssemblyName>
+    <AssemblyName>RapidCore.Redis.FunctionalTest</AssemblyName>
+    <RootNamespace>RapidCore.Redis.FunctionalTest</RootNamespace>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="Fakeiteasy" Version="3.3.2" />
+    <PackageReference Include="FakeItEasy" Version="4.0.0" />
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="1.1.1" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.0.0" />
     <PackageReference Include="xunit" Version="2.2.0" />
     <PackageReference Include="xunit.runner.visualstudio" Version="2.2.0" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\..\src\rapidcore.redis.csproj" />
+    <ProjectReference Include="..\main\rapidcore.redis.csproj" />
   </ItemGroup>
 </Project>
